refactor: move to types/ directory
