+++ packages/app/lib/internal/logger.ts
@@ -15,20 +15,9 @@
  *
  */
 
-export type LogLevelString = 'debug' | 'verbose' | 'info' | 'warn' | 'error' | 'silent';
+import type { LogCallback, LogOptions } from '../types/app';
 
-export interface LogOptions {
-  level?: LogLevelString;
-}
-
-export type LogCallback = (callbackParams: LogCallbackParams) => void;
-
-export interface LogCallbackParams {
-  level: LogLevelString;
-  message: string;
-  args: unknown[];
-  type: string;
-}
+type LogLevelString = 'debug' | 'verbose' | 'info' | 'warn' | 'error' | 'silent';
 
 /**
  * The JS SDK supports 5 log levels and also allows a user the ability to
