+++ packages/app/lib/types/app.ts
@@ -626,62 +626,3 @@ export interface LogOptions {
 
 // Re-export FirebaseApp as a named type for easier importing
 export type FirebaseApp = ReactNativeFirebase.FirebaseApp;
-
-/**
- * Internal namespace types for Firebase module registry
- */
-
-/**
- * Type for a Firebase module getter function that can optionally accept
- * a custom URL/region/databaseId parameter
- */
-export type ModuleGetter = {
-  (customUrlOrRegionOrDatabaseId?: string): ReactNativeFirebase.FirebaseModule;
-  [key: string]: unknown;
-};
-
-/**
- * Type for Firebase root object with module getters
- */
-export interface FirebaseRoot {
-  initializeApp: (
-    options: ReactNativeFirebase.FirebaseAppOptions,
-    configOrName?: string | ReactNativeFirebase.FirebaseAppConfig,
-  ) => Promise<ReactNativeFirebase.FirebaseApp>;
-  setReactNativeAsyncStorage: (asyncStorage: ReactNativeFirebase.ReactNativeAsyncStorage) => void;
-  app: (name?: string) => ReactNativeFirebase.FirebaseApp;
-  apps: ReactNativeFirebase.FirebaseApp[];
-  SDK_VERSION: string;
-  setLogLevel: (logLevel: ReactNativeFirebase.LogLevelString) => void;
-  [key: string]: unknown;
-}
-
-/**
- * Configuration interface for module namespace registration
- */
-export interface ModuleConfig {
-  namespace: string;
-  nativeModuleName?: string | string[];
-  hasMultiAppSupport?: boolean;
-  hasCustomUrlOrRegionSupport?: boolean;
-  nativeEvents?: boolean | string[];
-  disablePrependCustomUrlOrRegion?: boolean;
-  turboModule?: boolean;
-}
-
-/**
- * Extended configuration for namespace registration including native module details
- */
-export interface NamespaceConfig extends ModuleConfig {
-  nativeModuleName: string | string[];
-  nativeEvents: boolean | string[];
-  // ModuleClass can be FirebaseModule or any subclass of it
-  // Uses FirebaseAppBase (the concrete class type) rather than FirebaseApp (the augmented interface)
-  ModuleClass: new (
-    app: ReactNativeFirebase.FirebaseAppBase,
-    config: ModuleConfig,
-    customUrlOrRegion?: string | null,
-  ) => ReactNativeFirebase.FirebaseModule;
-  statics?: object;
-  version?: string;
-}
